branch changes from mitchel.jon (2nd)
- Add the cursor position relative to the dragged item when stating a
drag operation to the DragInfo object. The allows me to position the
default drag adorner under the cursor correctly.
- Added the ability to explicitly ignore drag and drop for a particular
control in visual tree. e.g. If you have a TextBox in a list control
that supports drag and drop, you can't do a drag selection of the text
in the TextBox. If you explicitly ignore it then drag selection will
work correctly.
- Fixed cast exception when attempting to start a drag operation from a
FrameworkContentElement. Need to use its parent as it can't be cast as a
UIElement.